lazy loading to solve flicker issue
